feat: split collection to a get/set/has trait; used mb_* lower methods
Refs #7
